+++ src/modules/dom.d.ts
@@ -1,0 +1,3 @@
+export function read(fn: () => void): number;
+export function write(fn: () => void): number;
+export function clear(id: number): void;